Accelerometer DRDY works on INT1/PE4/EXTI4 line
This means we are able to synchronise data reads from accelerometer
to when the data is ready on the LSM303DLHC sensor.

It is still unclear how to get the accelerometer and magnetometer
DRDY interrupts independently and simultaneuously but that's a work
in progress.

lsm303dlhc has been moved into the fcb_driver directory because
it required modifications.